web/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psutil==5.9.6
requests==2.31.0
PyJWT==2.8.0
